requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dacite==1.8.1
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.3
h11==0.14.0
idna==3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0